requirements.txt
# Django and core dependencies
Django==4.2.7
django-environ==0.11.2
Pillow>=10.0.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter
dj-database-url==2.1.0

# Static files and compression
whitenoise==6.6.0
django-compressor==4.4
rcssmin==1.1.1
rjsmin==1.2.2

# Forms and UI
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# API and CORS
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Image processing and uploads
Pillow>=10.0.0

# External API requests
requests==2.31.0

# Development and debugging
django-debug-toolbar==4.2.0

# Production server
gunicorn==21.2.0

# Caching
redis==5.0.1
django-redis==5.4.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Environment and configuration
python-decouple==3.8
